:feat :sparkles: created some important api

+++ src/app/modules/category/category.service.ts
@@ -4,17 +4,18 @@ import { ICategory } from './category.interface'
 import { Category } from './category.model'
 import unlinkFile from '../../../shared/unlinkFile'
 import { Bookmark } from '../bookmark/bookmark.model'
+import mongoose from 'mongoose'
 
 const createCategoryToDB = async (payload: ICategory) => {
-  const {name, image} = payload;
-  const isExistName = await Category.findOne({name: name})
+  const { name, image } = payload;
+  const isExistName = await Category.findOne({ name: name })
 
-  if(isExistName){
+  if (isExistName) {
     unlinkFile(image);
     throw new ApiError(StatusCodes.NOT_ACCEPTABLE, "This Category Name Already Exist");
   }
 
-  const createCategory:any = await Category.create(payload)
+  const createCategory: any = await Category.create(payload)
   if (!createCategory) {
     unlinkFile(image);
     throw new ApiError(StatusCodes.BAD_REQUEST, 'Failed to create Category')
@@ -29,36 +30,45 @@ const getCategoriesFromDB = async (): Promise<ICategory[]> => {
 }
 
 const updateCategoryToDB = async (id: string, payload: ICategory) => {
-    const isExistCategory:any = await Category.findById(id);
 
-    if(!isExistCategory){
-        throw new ApiError(StatusCodes.BAD_REQUEST, "Category doesn't exist");
-    }
-    
-    if (payload.image) {
-        unlinkFile(isExistCategory?.image);
-    }
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    throw new ApiError(StatusCodes.BAD_REQUEST, "Invalid Category ID")
+  }
+  const isExistCategory: any = await Category.findById(id);
 
-    const updateCategory = await Category.findOneAndUpdate(
-        { _id: id }, 
-        payload, 
-        { new: true}
-    )
+  if (!isExistCategory) {
+    throw new ApiError(StatusCodes.BAD_REQUEST, "Category doesn't exist");
+  }
 
-    return updateCategory
+  if (payload.image) {
+    unlinkFile(isExistCategory?.image);
+  }
+
+  const updateCategory = await Category.findOneAndUpdate(
+    { _id: id },
+    payload,
+    { new: true }
+  )
+
+  return updateCategory
 }
 
 const deleteCategoryToDB = async (id: string): Promise<ICategory | null> => {
-    const deleteCategory = await Category.findByIdAndDelete(id)
-    if (!deleteCategory) {
-        throw new ApiError(StatusCodes.BAD_REQUEST, "Category doesn't exist")
-    }
-    return deleteCategory
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    throw new ApiError(StatusCodes.BAD_REQUEST, "Invalid Category ID")
+  }
+
+  const deleteCategory = await Category.findByIdAndDelete(id)
+  if (!deleteCategory) {
+    throw new ApiError(StatusCodes.BAD_REQUEST, "Category doesn't exist")
+  }
+  return deleteCategory
 }
 
 export const CategoryService = {
-    createCategoryToDB,
-    getCategoriesFromDB,
-    updateCategoryToDB,
-    deleteCategoryToDB
+  createCategoryToDB,
+  getCategoriesFromDB,
+  updateCategoryToDB,
+  deleteCategoryToDB
 }
